Adds abstraction before adding mutual funds
