+++ src/server.js
@@ -1,24 +1,38 @@
 const puppeteer = require("puppeteer");
 const fs = require("fs/promises");
 
-const { aum_report_url } = require("./constants");
-const { scrapeCommonEntities, scrapeFunds } = require("./scraper-functions");
+const { aum_report_url, payout_url } = require("./constants");
+const {
+  scrapeCommonEntities,
+  scrapeFunds,
+  scrapeFundTypes,
+} = require("./scraper-functions");
 
 const startScraping = async () => {
   const browser = await puppeteer.launch();
   const page = await browser.newPage();
   await page.goto(aum_report_url);
-  const { fundTypesTable, amcsTable, fundCategoriesTable } =
-    await scrapeCommonEntities({ page, browser });
+  const fundTypes = await scrapeFundTypes({ page });
+  const { amcsTable, fundCategoriesTable } = await scrapeCommonEntities({
+    page,
+    browser,
+    linksToScrape: fundTypes?.map(({ link }) => link),
+  });
   const fundsTable = await scrapeFunds({
     page,
     browser,
-    linksToScrape: fundTypesTable?.map(({ link }) => link),
+    linksToScrape: fundTypes?.map(({ link }) => link),
   });
+
+  // start scraping for payouts
+  await page.goto(payout_url);
+  const payoutFundTypes = await scrapeFundTypes({ page });
+
+  console.log("Payout funds type", payoutFundTypes);
 
   await browser.close();
 
-  await fs.writeFile("fund-types.txt", JSON.stringify(fundTypesTable));
+  await fs.writeFile("fund-types.txt", JSON.stringify(fundTypes));
   await fs.writeFile("amcs.txt", JSON.stringify(amcsTable));
   await fs.writeFile(
     "fund-categories.txt",
